feature: compactable instances // fix circular reference between Compactable
